Rework SideBar
